feat: Add content moderation utils package
This commit adds a new package, utils, to the content moderation module. This package contains helper functions that can be used by the contentModerationHandler.go file and other files in the module. This will make the code more modular and easier to maintain in the future. Additionally, the contentModerationHandler.go file has been modified to use some of these new utility functions.